pyproject.toml
[project]
name = "dexter-eng"
version = "0.1.0"
description = "IA verticalizada para engenharia"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "typer>=0.9",
    "openai>=1.0",
    "pymupdf>=1.23",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
dexter = "dexter_eng.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
pythonpath = ["src"]
